fix: 404(2)

+++ controllers/cards.js
@@ -21,19 +21,18 @@ module.exports.createCard = (req, res) => {
 
 module.exports.deleteCard = (req, res) => {
   Card.findById({ _id: req.params.cardId })
-    .orFail(new Error('NotFound'))
-    .then(() => {
-      Card.findByIdAndRemove(req.params.cardId)
+    .then((card) => {
+      if (!card) {
+        return res.status(NOT_FOUND).send({ message: 'Карточка с указанным _id не найдена.' });
+      }
+      return Card.findByIdAndRemove(req.params.cardId)
         .then(() => {
-          res.send({ _id: req.params.cardId });
+          res.status(200).send({ _id: req.params.cardId });
         });
     })
     .catch((err) => {
       if (err.kind === 'ObjectId') {
         return res.status(BAD_REQUEST).send({ message: 'Некорректный формат id.' });
-      }
-      if (err.messsage === 'NotFound') {
-        return res.status(NOT_FOUND).send({ message: 'Передан несуществующий _id карточки.' });
       }
       return res.status(INTERNAL_SERVER_ERROR).send({ message: 'Ошибка по умолчанию.' });
     });
@@ -45,16 +44,15 @@ module.exports.addLike = (req, res) => {
     { $addToSet: { likes: req.user._id } },
     { new: true },
   )
-    .orFail(new Error('NotFound'))
-    .then(() => {
-      res.send({ _id: req.params.cardId });
+    .then((card) => {
+      if (!card) {
+        return res.status(NOT_FOUND).send({ message: 'Передан несуществующий _id карточки.' });
+      }
+      return res.status(200).send({ _id: req.params.cardId });
     })
     .catch((err) => {
       if (err.name === 'ValidationError') {
         return res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные для постановки/снятии лайка.' });
-      }
-      if (err.messsage === 'NotFound') {
-        return res.status(NOT_FOUND).send({ message: 'Передан несуществующий _id карточки.' });
       }
       if (err.kind === 'ObjectId') {
         return res.status(BAD_REQUEST).send({ message: 'Некорректный формат id.' });
@@ -69,16 +67,15 @@ module.exports.removeLike = (req, res) => {
     { $pull: { likes: req.user._id } },
     { new: true },
   )
-    .orFail(new Error('NotFound'))
-    .then(() => {
-      res.send({ _id: req.params.cardId });
+    .then((card) => {
+      if (!card) {
+        return res.status(NOT_FOUND).send({ message: 'Передан несуществующий _id карточки.' });
+      }
+      return res.status(200).send({ _id: req.params.cardId });
     })
     .catch((err) => {
       if (err.name === 'ValidationError') {
         return res.status(BAD_REQUEST).send({ message: 'Переданы некорректные данные для постановки/снятии лайка.' });
-      }
-      if (err.messsage === 'NotFound') {
-        return res.status(NOT_FOUND).send({ message: 'Передан несуществующий _id карточки.' });
       }
       if (err.kind === 'ObjectId') {
         return res.status(BAD_REQUEST).send({ message: 'Некорректный формат id.' });
@@ -86,3 +83,4 @@ module.exports.removeLike = (req, res) => {
       return res.status(INTERNAL_SERVER_ERROR).send({ message: 'Ошибка по умолчанию.' });
     });
 };
+
